Extract CleanupAfter, CleanupBefore, CleanupTest
Further modularisation to reduce coupling.
